requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu
rouge_score
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
torch
notebook
boto3
mypy-boto3
python-box==6.0.2
ensure==1.0.2
uvicorn==0.18.3
Jinja2

# -e . #installs the setup.py